Cache ToneJS script in indexedDB
ToneJS needs to be dynamically loaded on user interaction,
but we don't want to wait for a download after we try to
start audio (it should be responsive).
But, we can't import straight away, because that triggers
the AudioContext warning in the browser.
So we need to download the script to have it ready, then
import it on user interaction.
We also cache the script text using indexedDB.