Change note transition behaviour
Notes now longer overlap when transitioning. A promise
ensures correct stopping of each note before a new one
begins. Because player.fadeOut can't be changed dynamically,
the volume parameter is automated with a ramp to get
different fadeout times - fast for transitioning between
notes, and slow for a nice fadeout on stopping.